+++ examples/libs/aph-math/index.ts
@@ -4,9 +4,11 @@ import { QuadTree, QuadTreeItem } from './structs-spatial/quad-tree';
 import { PathContext, PathSegment, Path } from './pathfinding/path';
 import { PathFinderContext, PathFinder, BreadthFirstSearch, Dijkstra, AStarSearch } from './pathfinding/pathfinding';
 import { GridMap, MAP_TYPE_TILE, MAP_TYPE_OCTILE } from './pathfinding/gridmap';
+import * as Steering from './steering';
 
 export {
   Random,
+  Steering,
   QuadTree, QuadTreeItem, PerlinNoise,
   PathContext, PathSegment, Path,
   PathFinderContext, PathFinder, BreadthFirstSearch, Dijkstra, AStarSearch,