+++ examples/src/02-pixi-intro/text.ts
@@ -3,15 +3,36 @@ import { PIXIExample } from '../utils/APHExample';
 
 export class Text extends PIXIExample {
 
+    text: PIXI.Text;
+
 	load() {
-        // TODO put your code here
-        // 1) define PIXI.TextStyle
-        // 2) create PIXI.Text object
-        // 3) add the text to the stage
-        // 4) in update method, animate the text from left to right
+        const style = new PIXI.TextStyle({
+            fontFamily: 'Arial',
+            fontSize: 36,
+            fill: ['#ffffff', '#00ff99'],
+            stroke: '#4a1850',
+            strokeThickness: 5
+        });
+
+        this.text = new PIXI.Text('Hello World', style);
+        this.text.position.set(this.app.screen.width / 2, this.app.screen.height / 2);
+        this.text.anchor.set(0.5);
+        this.app.stage.addChild(this.text);
     }
 
+    directionLeft = false;
+
     update(deltaTime: number) {
-        // TODO put your code here
+        if(this.directionLeft) {
+            this.text.position.x -= deltaTime * 2;
+        } else {
+            this.text.position.x += deltaTime * 2;
+        }
+
+        if(this.text.getBounds().left < 0) {
+            this.directionLeft = false;
+        } else if(this.text.getBounds().right > this.app.screen.width) {
+            this.directionLeft = true;
+        }
     }
 }