+++ examples/src/examples.ts
@@ -16,6 +16,20 @@ import { DistributionNormal, DistributionUniform } from './04-space/distribution
 import { Pathfinding } from './04-space/pathfinding';
 import { PerlinNoiseView as PerlinNoise } from './04-space/perlin-noise';
 import { QuadTreeView as QuadTree } from './04-space/quadtree';
+import { Missile } from './05-dynamics/missile';
+import { Rotation } from './05-dynamics/rotation';
+import { SteeringCombined } from './05-dynamics/steering-combined'; 
+import { SteeringSeek } from './05-dynamics/steering-seek'; 
+import { SteeringEvade } from './05-dynamics/steering-evade'; 
+import { SteeringPursuit } from './05-dynamics/steering-pursuit'; 
+import { SteeringWander } from './05-dynamics/steering-wander'; 
+import { MatterAirFriction } from './06-physics/matter-airfriction'; 
+import { MatterFriction } from './06-physics/matter-friction'; 
+import { MatterBridge } from './06-physics/matter-bridge'; 
+import { MatterCar } from './06-physics/matter-car'; 
+import { MatterChains } from './06-physics/matter-chains'; 
+import { MatterCradle } from './06-physics/matter-cradle'; 
+import { MatterMixedShapes } from './06-physics/matter-mixed-shapes'; 
 
 import { PixiHitTest } from './others/pixi-hittest';
 
@@ -39,5 +53,19 @@ import { PixiHitTest } from './others/pixi-hittest';
     Pathfinding,
     PerlinNoise,
     QuadTree,
+    Missile,
+    Rotation,
+    SteeringCombined,
+    SteeringSeek,
+    SteeringEvade,
+    SteeringPursuit,
+    SteeringWander,
+    MatterAirFriction,
+    MatterFriction,
+    MatterBridge,
+    MatterCar,
+    MatterChains,
+    MatterCradle,
+    MatterMixedShapes,
     PixiHitTest
 };