+++ wiki/sidebars.js
@@ -38,7 +38,14 @@ module.exports = {
 			items: [{
 				label: 'Pixi Introduction',
 				type: 'category',
-				items: ['examples/pixi-intro/config', 'examples/pixi-intro/animation']
+				items: ['examples/pixi-intro/config', 'examples/pixi-intro/animation', 'examples/pixi-intro/button', 
+				'examples/pixi-intro/loop', 'examples/pixi-intro/particles', 'examples/pixi-intro/primitives', 
+				'examples/pixi-intro/sound', 'examples/pixi-intro/text']
+			},
+			{
+				label: 'ThreeJS Introduction',
+				type: 'category',
+				items: ['examples/three-intro/shaders', 'examples/three-intro/scene', 'examples/three-intro/sprites']
 			},
 			{
 				label: 'Components',
@@ -48,7 +55,22 @@ module.exports = {
 			{
 				label: 'Space',
 				type: 'category',
-				items: ['examples/space/random-dist', 'examples/space/pathfinding', 'examples/space/perlin', 'examples/space/quadtree']
+				items: ['examples/space/random-dist', 'examples/space/pathfinding', 'examples/space/perlin', 
+				'examples/space/quadtree']
+			},
+			{
+				label: 'Dynamics',
+				type: 'category',
+				items: ['examples/dynamics/missile', 'examples/dynamics/rotation', 'examples/dynamics/steering-seek', 
+				'examples/dynamics/steering-evade', 'examples/dynamics/steering-pursuit', 'examples/dynamics/steering-wander',
+				'examples/dynamics/steering-combined']
+			},
+			{
+				label: 'Physics',
+				type: 'category',
+				items: ['examples/physics/matter-friction', 'examples/physics/matter-airfriction',
+				'examples/physics/matter-bridge', 'examples/physics/matter-car', 'examples/physics/matter-chains',
+				'examples/physics/matter-cradle', 'examples/physics/matter-mixed-shapes']
 			},
 			{
 				label: 'Other',
@@ -64,7 +86,7 @@ module.exports = {
 		{
 			label: 'Tutorials',
 			type: 'category',
-			items: ['tutorials/pixi_intro', 'tutorials/pixi_positions', 'tutorials/three_intro']
+			items: ['tutorials/pixi_intro', 'tutorials/pixi_positions', 'tutorials/three_intro', 'tutorials/matterjs']
 		},
 	].filter(Boolean)
 }