+++ wiki/src/internals/aph-colors.js
@@ -1,7 +1,11 @@
 module.exports = {
-	midnight: '#03060c',
-	midnightLight: '#121824',
+	midnight: '#121824',
+	midnightLight: '#3a414e',
+	midnightDark: '#03060c',
 	white: '#ffffff',
+	goldenrodLight: '#ffd179',
+	goldenrod: '#fbb227',
+	goldenrodDark: '#a96f00',
 	tomatoLight: '#ff867f',
 	tomato: '#cf2f25',
 	tomatoDark: '#7d0600',